# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API for Big Bad Con",
    "version": "1.0",
    "title": "Big Bad Con data service REST API",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Jeremy Tidwell",
      "url": "https://www.bigbadcon.com/",
      "email": "necanthrope@gmail.com"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8091",
  "basePath": "/",
...
  "paths": {
...
    "/events/all": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "List all events for the current calendar year",
        "operationId": "findAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/all/public": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "A public facing list of all events for the current calendar year. Does not require login.",
        "operationId": "findAllPublicUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventsPublic"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/category/{category}": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "Find all events for a category.",
        "operationId": "findByCategoryUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "description": "category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/count": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "Count of events for the current calendar year",
        "operationId": "getCountUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/create": {
      "put": {
        "tags": [
          "events-controller"
        ],
        "summary": "Create an event for the current year.",
        "operationId": "createEventUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "eventCreate",
            "description": "eventCreate",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateEventDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/eventmeta": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "List all possible categories and tags for events.",
        "operationId": "findMetaTypesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventMetadataDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/favorite/create": {
      "post": {
        "tags": [
          "events-controller"
        ],
        "summary": "Create a favorite/starred event for another user by event ID and user ID",
        "description": "Admin only.",
        "operationId": "createUserEventFavoriteUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "eventAndUser",
            "description": "eventAndUser",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventAndUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventAndUser"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/favorite/delete": {
      "delete": {
        "tags": [
          "events-controller"
        ],
        "summary": "Remove a favorite/starred event for another user by event ID and user ID",
        "description": "Admin only.",
        "operationId": "deleteUserEventFavoriteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "eventAndUser",
            "description": "eventAndUser",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventAndUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventAndUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/events/find": {
      "post": {
        "tags": [
          "events-controller"
        ],
        "summary": "Find an event by event ID.",
        "operationId": "findOneUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "eventFind",
            "description": "eventFind",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventFind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/me": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "Find all events for the current user.",
        "operationId": "findForMeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/me/favorite/create": {
      "post": {
        "tags": [
          "events-controller"
        ],
        "summary": "Create a favorite/starred favoriteEvent for myself by favoriteEvent ID",
        "operationId": "createMyUserEventFavoriteUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "favoriteEvent",
            "description": "favoriteEvent",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FavoriteEvent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FavoriteEvent"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/me/favorite/delete": {
      "delete": {
        "tags": [
          "events-controller"
        ],
        "summary": "Delete a favorite/starred favoriteEvent for myself by favoriteEvent ID",
        "operationId": "deleteMyUserEventFavoriteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "favoriteEvent",
            "description": "favoriteEvent",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FavoriteEvent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FavoriteEvent"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/events/me/favorites": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "Find all favorited events for the current user.",
        "operationId": "findFavoritesForMeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Events"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/page/public/{length}/{offset}": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "List a page events for the current calendar year",
        "operationId": "findPagePublicUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "length",
            "in": "path",
            "description": "length",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "offset",
            "in": "path",
            "description": "offset",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventsPublic"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/page/{length}/{offset}": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "List a page events for the current calendar year",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "length",
            "in": "path",
            "description": "length",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "offset",
            "in": "path",
            "description": "offset",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/since/{epochtime}": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "Find all events for the current year, updated after a time specified as UTC epoch.",
        "operationId": "findUpdatedAfterEpochTimeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "epochtime",
            "in": "path",
            "description": "epochtime",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/since/{epochtime}/public": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "A public facing list of all events for the current calendar year since a provided epoch time. Does not require login.",
        "operationId": "findAllSincePublicUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "epochtime",
            "in": "path",
            "description": "epochtime",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventsPublic"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/user": {
      "post": {
        "tags": [
          "events-controller"
        ],
        "summary": "Find all events for a registered user by user ID",
        "description": "Admin only.",
        "operationId": "findForUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "eventFind",
            "description": "eventFind",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventFind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/events/year/{year}/category/{category}": {
      "get": {
        "tags": [
          "events-for-year-controller"
        ],
        "summary": "Find all events for a given year, for a category.",
        "operationId": "findByCategoryUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "year",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "category",
            "in": "path",
            "description": "category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EventsPublic": {
      "type": "object",
      "properties": {
        "blogId": {
          "type": "integer",
          "format": "int64"
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BbcEventCategoriesShort"
          }
        },
        "eventAttributes": {
          "type": "string"
        },
        "eventCategoryId": {
          "type": "integer",
          "format": "int64"
        },
        "eventEndDate": {
          "type": "string",
          "format": "date"
        },
        "eventEndTime": {
          "$ref": "#/definitions/Time"
        },
        "eventId": {
          "type": "integer",
          "format": "int64"
        },
        "eventName": {
          "type": "string"
        },
        "eventSlug": {
          "type": "string"
        },
        "eventStartDate": {
          "type": "string",
          "format": "date"
        },
        "eventStartTime": {
          "$ref": "#/definitions/Time"
        },
        "eventStatus": {
          "type": "integer",
          "format": "int32"
        },
        "groupId": {
          "type": "integer",
          "format": "int64"
        },
        "lastUpdated": {
          "$ref": "#/definitions/Timestamp"
        },
        "metadata": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Postmeta"
          }
        },
        "postContent": {
          "type": "string"
        },
        "postId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "EventMetadataDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
...
    "FavoriteEvent": {
      "type": "object",
      "properties": {
        "eventId": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "EventFind": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "EventAndUser": {
      "type": "object",
      "properties": {
        "eventId": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
  }
}